Make Rodauth Hanami feature configurable, move into separate file
