+++ lib/rodauth/hanami.rb
@@ -1,0 +1,52 @@
+# frozen_string_literal: true
+
+require "rodauth"
+
+Rodauth::Feature.define(:hanami) do
+  auth_value_method :hanami_base_view_class, nil
+
+  # Renders templates with layout.
+  def view(template, title)
+    return super unless view_template?(template)
+
+    view_rendering.template(
+      base_view.class.layout_path(base_view.config.layout),
+      view_rendering.scope(rodauth: self)
+    ) { render(template) }
+  end
+
+  # Renders templates without layout.
+  def render(template)
+    return super unless view_template?(template)
+
+    view_rendering.template(
+      view_template_name(template),
+      view_rendering.scope(rodauth: self)
+    )
+  end
+
+  private
+
+  def view_template?(template_name)
+    view_rendering.renderer.send(
+      :lookup,
+      view_template_name(template_name),
+      base_view.config.default_format
+    )
+  end
+
+  def view_template_name(template_name)
+    "authentication_app/#{template_name.tr("-", "_")}"
+  end
+
+  def view_rendering
+    @view_rendering ||= base_view.rendering(
+      format: base_view.config.default_format,
+      context: base_view.config.default_context
+    )
+  end
+
+  def base_view
+    @base_view ||= Class.new(hanami_base_view_class.call).new
+  end
+end
